Refactor createInfoNode() into its own file
