+++ plugins/swagger-to-graphql-source/src/swaggerGraphQLNodesFactories.js
@@ -1,38 +1,5 @@
-const INFO_NODE_TYPE = "SwaggerOpenApiInfo";
 const PATH_NODE_TYPE = "SwaggerOpenApiPath";
 const DEFINITION_NODE_TYPE = "SwaggerOpenApiDefinition";
-
-exports.createInfoNode = props => {
-  const {
-    api,
-    parentFile,
-    gatsbyNodeApi: { createNodeId, createContentDigest },
-  } = props
-
-  const { info, swagger, host, basePath, schemes, consumes, produces, tags } = api;
-
-  const { name, id: parentId } = parentFile;
-
-  return {
-    ...info,
-    swagger,
-    host,
-    basePath,
-    schemes,
-    consumes,
-    produces,
-    tags,
-    id: createNodeId(`${name} ${info.title} ${info.version}`),
-    name: name,
-    children: [],
-    internal: {
-      content: "",
-      contentDigest: createContentDigest(info),
-      type: INFO_NODE_TYPE,
-    },
-    parent: parentId,
-  }
-}
 
 exports.createPathNodes = props => {
   const {
